Move utils to seprate package.
